Make generated migration scripts more robust if queueing one migration fails
Customers often use migration scripts generated with `gh gei generate-script` to queue up many migrations.

If queueing up one migration fails, then at the moment, the entire script explodes. This is a pretty painful experience, since customers will often leave the script running unattended (e.g. overnight), so they won't know when it terminates and they won't be able to intervene.

This changes the behaviour of our generated scripts so that, if one migration cannot be queued, we print out an error and move on to the next one.

If we have failed to queue up a migration, if applicable, it will be included in the "failed migrations" count printed at the end of the script and we will not attempt to download logs if the `--download-migration-logs` option was selected.

Fixes https://github.com/github/gh-gei/issues/781.