Including the repo name in Bitbucket Server migration archive filenames
When running a Bitbucket Server migration, the archive from the Bitbucket Server instance is stored in the `bbs_archive_downloads` directory. At the moment, the filename only contains the migration ID.

This makes it hard to identify which file is which, which makes life difficult when we want to ask a customer for a copy of their archive for debugging purposes.

This updates the download process so we include the migration ID and the BBS repo name in the filename of the downloaded TAR files.