Support GHAE as source (#846)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Migrating from GHAE follows the same path as migrating from GHES
(generate archive, upload to blob storage, etc). As part of this path we
check the GHES version in order to determine whether we need blob
storage args or not (GHES v3.8+ handles the blob storage for us).

However, when we do the version check API call against GHAE instead of
returning a string like "3.8.0" it returns "GitHub AE" and our version
check code crashes trying to parse it. This PR fixes that by having it
treat any non-parseable value as < 3.8.0.

Fixes #832 

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [ ] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

---------

Co-authored-by: Tim Rogers <timrogers@github.com>