Consistently convert GraphQL errors to `OctoshiftCliException`s with helpful error messages across `GithubApi`
In `GithubApi`, we have the helpful `EnsureSuccessGraphQLResponse`
helper which checks the GraphQL response for errors, and if an error
is found, it raises a `OctoshiftCliException`.

Currently, this helper is not used consistently across GraphQL API
calls, which means that we have some cases where the GraphQL API
returns an error and the CLI crashes with an unhelpful
`System.InvalidOperationException: Cannot access child value on Newtonsoft.Json.Linq.JValue.`.

To find out what actually went wrong, the user has to either:

* (a) use the `--verbose` option or
* (b) manually check the verbose log file

This consistently applies the `EnsureSuccessGraphQLResponse` helper so
that helpful exceptions with helpful messages are raised.

I do not apply this to `ReclaimMannequin` (which has its own error
handling) and `GetMannequins` (which has some complex pagination logic
which I didn't want to mess with right now).

Fixes https://github.com/github/gh-gei/issues/867.