Improve error message when `CreateMigrationSource` is called with insufficient permissions
When using the `migrate-repo` command, under the hood, we call the
`CreateMigrationSource` GraphQL mutation. If your target PAT doesn't
have the correct permissions, the API returns an error like this:

> timrogers does not have the correct permissions to execute `CreateMigrationSource`

This error isn't very helpful, as it doesn't tell you what you should do
to fix it. It would be great to have a more instructive error message
that guides the customer on where they've gone wrong.

It isn't trivial for us to change the actual error returned by the API
because it is part of GitHub's shared GraphQL logic. This takes a
different approach: catching the specific error and decorating the
message with some more helpful guidance on what to do next.