Continue execution of `gh gei reclaim-mannequin --csv` if a username doesn't exist
Since #868, the `gh gei reclaim-mannequin` command with the `--csv`
option now exits if it is given a username that doesn't exist, rather
than printing a warning and continuing to the next mapping in the CSV.

This fixes that behaviour, handling the `OctoshiftCliException` thrown
by `GithubApi` and printing a warning instead of exiting.

Fixes https://github.com/github/gh-gei/issues/897.