Validate `bbs2gh` env vars in the generated script (#923)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Implements part of #629 

Added some powershell to the start of the script generated by `bbs2gh
generate-script` that validates the proper environment variables are set
before executing the rest of the script.

This one was a bit trickier than `ado2gh` because depending on what args
you pass to `generate-script` the script will expect different sets of
env vars. This was the pseudocode I came up with before coding this up:
```
always validate GH_PAT
always validate BBS_PASSWORD unless args.Kerberos
if args.BbsUsername.IsNullOrWhitespace() and !args.Kerberos validate BBS_USERNAME
if args.AwsBucketName.HasValue() or args.AwsRegion.HasValue()
	Validate AWS_ACCESS_KEY_ID
	Validate AWS_SECRET_ACCESS_KEY
if AwsBucketName.IsNull and AwsRegion.IsNull
	Validate AZURE_STORAGE_CONNECTION_STRING
if SmbUser.HasValue
	Validate SMB_PASSWORD
```

These are all the various validation messages:
```
if (-not $env:GH_PAT) {
    Write-Error "GH_PAT environment variable must be set to a valid GitHub Personal Access Token with the appropriate scopes. For more information see https://docs.github.com/en/migrations/using-github-enterprise-importer/preparing-to-migrate-with-github-enterprise-importer/managing-access-for-github-enterprise-importer#creating-a-personal-access-token-for-github-enterprise-importer"
    exit 1
} else {
    Write-Host "GH_PAT environment variable is set and will be used to authenticate to GitHub."
}

if (-not $env:BBS_USERNAME) {
    Write-Error "BBS_USERNAME environment variable must be set to a valid user that will be used to call BBS API's to generate a migration archive."
    exit 1
} else {
    Write-Host "BBS_USERNAME environment variable is set and will be used to authenticate to BBS APIs."
}

if (-not $env:BBS_PASSWORD) {
    Write-Error "BBS_PASSWORD environment variable must be set to a valid password that will be used to call BBS API's to generate a migration archive."
    exit 1
} else {
    Write-Host "BBS_PASSWORD environment variable is set and will be used to authenticate to BBS APIs."
}

if (-not $env:AZURE_STORAGE_CONNECTION_STRING) {
    Write-Error "AZURE_STORAGE_CONNECTION_STRING environment variable must be set to a valid Azure Storage Connection String that will be used to upload the migration archive to Azure Blob Storage."
    exit 1
} else {
    Write-Host "AZURE_STORAGE_CONNECTION_STRING environment variable is set and will be used to upload the migration archive to Azure Blob Storage."
}

if (-not $env:AWS_ACCESS_KEY_ID) {
    Write-Error "AWS_ACCESS_KEY_ID environment variable must be set to a valid AWS Access Key ID that will be used to upload the migration archive to AWS S3."
    exit 1
} else {
    Write-Host "AWS_ACCESS_KEY_ID environment variable is set and will be used to upload the migration archive to AWS S3."
}

if (-not $env:AWS_SECRET_ACCESS_KEY) {
    Write-Error "AWS_SECRET_ACCESS_KEY environment variable must be set to a valid AWS Secret Access Key that will be used to upload the migration archive to AWS S3."
    exit 1
} else {
    Write-Host "AWS_SECRET_ACCESS_KEY environment variable is set and will be used to upload the migration archive to AWS S3."
}

if (-not $env:SMB_PASSWORD) {
    Write-Error "SMB_PASSWORD environment variable must be set to a valid password that will be used to download the migration archive from your BBS server using SMB."
    exit 1
} else {
    Write-Host "SMB_PASSWORD environment variable is set and will be used to download the migration archive from your BBS server using SMB."
}
```

Note: I updated this PR to target another PR branch (instead of main) so
the diff is correct for reviewers. Will retarget it to main once that
other PR merges.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->