More robust retry logic on GETs (#928)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Fixes #726 

We actually already have retries when polling archive generation status.
But in the linked issue they don't get triggered because it's a
`System.TimeoutException` and the retry logic code is only triggering on
`HttpRequestException`. We originally did it this way so we can provide
an exception filter, and only retry on specific HTTP error codes.
However, as we've evolved our code we have a few places where we set
that filter to `_ => true` which we probably intended to mean retry on
any error - but what it actually is doing is retry on any HTTP error.

This PR updates all the places we were doing that to use
`RetryPolicy.Retry` instead of `HttpRetry` so it will retry on all
exception types.

Also updated `RetryPolicy.Retry` to no longer return a `PolicyResult`
object, and instead return the actual value (or throw an exception).
This makes it more convenient to use.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->