Validate gei env vars in script (#924)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Implements part of https://github.com/github/gh-gei/issues/629

Added some powershell to the start of the script generated by gei
generate-script that validates the proper environment variables are set
before executing the rest of the script. GH_PAT always needs to be set,
and if `--ghes-api-url` has a value then either
AZURE_STORAGE_CONNECTION_STRING or the AWS env vars need to be set.

I intentionally did not add any validation to the ADO migration scripts
as that functionality will be removed from `gei` soon.

With GHES source, whether we need to validate Azure/AWS env vars or not
depends on what version of GHES it is. We had an existing function in
`MigrateRepoCommandHandler` that would retrieve the GHES version number,
parse it, and determine whether we need blob storage creds or not.
Instead of copy-pasting that I refactored it out into a new service -
`GhesVersionCheckerService` - and injected it into both handlers (and
added a few unit tests for the new service).

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->