Proper URL Encoding (#927)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Fixes #772 

We had a bunch of places in the code where we were replacing spaces in
URL's with %20. The more robust way to do this is to use
`HttpUtility.UrlEncode()` which will also catch other special characters
(e.g. parentheses which are valid in ADO Team Project names).

I created an extension method `string.UrlEncode` to make the code a bit
more concise.

Everywhere we craft a URL with user-provided values we should be calling
`UrlEncode()` on the user provided values at the place in the code where
we stuff them into a URL. This is alot of places in our
AdoApi/GitApi/etc classes.

I updated some of our unit test sample data to have special characters,
and updated the expected values in a whole whack of existing tests to
give a good amount of coverage over the new behavior.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->