Wait as default (#929)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created) - #946 
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Implements #705 

Changing the default behavior of `migrate-repo` and `migrate-org` to be
the same as the `--wait` flag.

The `--wait` flag is now unnecessary, but for compat reasons I'm leaving
it in. I made it a hidden option, and if you pass it you'll get this
warning: `--wait flag is obsolete and will be removed in a future
version. The default behavior is now to wait.`

Added a new option `--queue-only` which will result in the same behavior
as the previous default behavior.

Also updating `generate-script` commands to include the `--queue-only`
flag for parallel scripts (and remove `--wait` from the sequential
scripts).

For customers that have pre-existing parallel scripts (without the
`--wait` flag) this will result in a change in behavior (scripts that
used to be parallel will now run sequentially), but it shouldn't
**break** their scripts (the integration tests confirm this). Anybody
that relies on the default behavior will start getting a warning that
says:
`[WARNING] The default behavior has changed from only queueing the
migration, to waiting for the migration to finish. If you ran this as
part of a script to run multiple migrations in parallel, consider using
the new --queue-only option to preserve the previous default behavior.`.
In a future release we'll remove this warning.

Any existing sequential scripts will continue to work just fine, they
will just see warnings about using an unnecessary `--wait` flag now.

Added validation that both `--wait` and `--queue-only` can't be passed
in together.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->