Add option to set target repo visibility (#953)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

Implements part of #5 (FINALLY!)

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

This was originally implemented by @tambling in #647 . However there was
some back and forth on some design questions and it never got merged,
and now that PR/branch is so far behind main I figured it would be
easier to just redo the changes in a new branch rather than trying to
resolve the merge/rebase conflicts with the previous PR/branch.

This adds a `--target-repo-visibility` argument to all 3 `migrate-repo`
commands. Octoshift backend already supports this arg in the API's, we
just never used it (until now).

I will do a separate PR that updates `generate-script`.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->