Add target-repo-visibility to generated scripts (#956)
<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [x] Docs updated (or issue created)
- [x] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

Closes #5 

Updated all 3 `generate-script` commands to always include the
`--target-repo-visibility` arg. For `gei` it will inspect the
source-repo and set the value appropriately in the script. For `ado2gh`
and `bbs2gh` it will always set the value to `private`.

Originally I added a separate function `GithubApi.GetRepoVisibility()`
but this resulted in an API call per repo when generating the script
which made it significantly slower for large orgs. The API we already
call to get the list of repos also returns the visibility, so I updated
`GithubApi.GetRepos()` to return a tuple with repo name and visibility.

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

---------

Co-authored-by: Arin Ghazarian <aringhazarian@github.com>