Improve logging when provided blob storage credentials are ignored in a GHES v3.8.0+ migration
When running a GitHub-to-GitHub migration with `gh gei migrate-repo`,
the CLI alllows you to pass in blob storage credentials.

These are only used for GHES-to-GHEC migrations where the GHES instance
is running a version before v3.8.0.

If you're running a GHEC-to-GHEC migration, no blob storage is involved,
and if you're migrating from GHES v3.8.0 onwards, then blob storage is
configured on the GHES Management Console, and values passed through
the CLI are not used.

This improves the CLI's logging to make it clearer that your supplied
blob storage credentials will not be used, making the message itself
easier to understand and changing it to a warning.

This will help to avoid users thinking that their supplied blob storage
credentials are being used, when in fact they are ignored.