Include the date with the timestamp when writing to the log (#963)
This PR will Include the date with the timestamp when writing to the
log.

Currently, GEI CLI writes logs with only timestamp as follows: 

```
[8:42] [INFO] You are running the latest version of the gei CLI [v0.40]
[8:42] [INFO] Migrating Repo...
[8:42] [INFO] GITHUB SOURCE ORG: <source org>
[8:42] [INFO] SOURCE REPO: <source repo>
[8:42] [INFO] GITHUB TARGET ORG: <target org>
[8:42] [INFO] TARGET REPO: <target repo>
```

I see the `octoshift.log` includes the date in the filename but it would
be good if the logs themselves include the date for troubleshooting.

Thanks!

- [x] Did you write/update appropriate tests -> No because it's small
update. But I tested the logic worked on my local.
- [x] Release notes updated (if appropriate)
- [ ] Appropriate logging output
- [ ] Issue linked
- [ ] Docs updated (or issue created)
- [ ] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

---------

Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>