Unhide the `--archive-download-host` option for `gh bbs2gh migrate-repo` and `generate-script`
When we introduced the `--archive-download-host` argument to customize
the host used for archive downloads in Bitbucket Server migrations, we
marked the option as hidden.

We made this decision because we didn't know how useful this option
would be and how many people would need it.

Based on feedback from numerous customers, I'm confident that there are
many people who need to customize this option. This unhides it the
`migrate-repo` and `generate-script` commands' documentation to make it
more discoverable.

Fixes https://github.com/github/gh-gei/issues/972.