Unhide the `--archive-download-host` option for `gh bbs2gh migrate-repo` and `generate-script` (#973)
When we introduced the `--archive-download-host` argument to customize
the host used for archive downloads in Bitbucket Server migrations, we
marked the option as hidden.

We made this decision because we didn't know how useful this option
would be and how many people would need it.

Based on feedback from numerous customers, I'm confident that there are
many people who need to customize this option. This unhides it the
`migrate-repo` and `generate-script` commands' documentation to make it
more discoverable.

Fixes https://github.com/github/gh-gei/issues/972.

<!--
For the checkboxes below you must check each one to indicate that you
either did the relevant task, or considered it and decided there was
nothing that needed doing
-->

- [x] Did you write/update appropriate tests
- [x] Release notes updated (if appropriate)
- [x] Appropriate logging output
- [x] Issue linked
- [ ] Docs updated (or issue created)
- [ ] New package licenses are added to `ThirdPartyNotices.txt` (if
applicable)

<!--
For docs we should review the docs at:

https://docs.github.com/en/early-access/github/migrating-with-github-enterprise-importer
and the README.md in this repo

If a doc update is required based on the changes in this PR, it is
sufficient to create an issue and link to it here. The doc update can be
made later/separately.

The process to update the docs can be found here:
https://github.com/github/docs-early-access#opening-prs

The markdown files are here: 

https://github.com/github/docs-early-access/tree/main/content/github/migrating-with-github-enterprise-importer
-->

---------

Co-authored-by: Dylan Smith <dylanfromwinnipeg@gmail.com>