Improve log secret sanitization to remove secrets that have been URL encoded
When logging with `Octologger` to the console output and log files,
we sanitize secrets, replacing them with `***`.

We identify what values are secrets - and so should be replaced - by
marking certain command line arguments as secret.

Each time we log, we go through those registered secret values and
perform a simple "search and replace" to remove them.

This updates that logic to also remove the URL-encoded version of
each secret which can appear where GraphQL request bodies are logged.