Include the `databaseId` (GUID) in the verbose logs when starting a migration
Right now, when starting a migration, we only print the GraphQL ID
(beginning `RM_` to the logs).

This improves the logs by also outputting the GUID (aka
`databaseId`), which is referred to in certain places (e.g. the
Migration Log issue and text file).

We don't actually write any new logging code to do this - we just
request the `databaseId` field in our GraphQL queries, and rely on
existing behaviour which logs GraphQL responses to the verbose log.